[TEST] get_info
in the test of cheapest_get_info there is a kind of Bug. The Scraper
works without an issue, but in unit testing, the tests got and error
that doesn't make sense, showing a 2 157.27 price, as an example, that
couldn't turn into a float, when all spaces where removed and the
scraper doesn't show this issue.

This is push beacuse of the scraper file situation (it works)