[FEAT] coin_dictionary
Instead of give a int that marks where to stract in get_price function,
now fives how the page manage the coin. In the case of Ebay, they use
the 'MXN $' string and not just '$' for the coin management. This
highlight could be usefull for a futher universal e-commerce scraper.