[BUG] [FEAT] get_price
In unit testing, the last version has a bug with the error riser in the
test. In the search of a way to prevent this, the coin_dictionary was
improve as the string extractor was changed with a replacement of the
way that the page manage the coins.

Instead of a:
 - price_string[coin_symnol:] call
 - price_string.replace(coin_symbol, '')